requirements.txt
# Jasper core dependencies
APScheduler
argparse
mock
pytz
PyYAML
requests

# Audio play and record
pyaudio
pymad

# APScheduler require futures in python 2.7
futures

# HN module
beautifulsoup4
semantic

# MPDControl module
#python-mpd
